Backend/Fast API/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0

# Database
pymongo==4.6.0

# AI & Embeddings
openai>=1.0.0
langchain==0.3.3
langchain-community==0.3.2
langchain-openai==0.2.2
tiktoken==0.7.0

# Vector Search
faiss-cpu==1.12.0

# ML Models
sentence-transformers==2.2.2

# Utils
python-dotenv==1.0.1
pydantic==2.5.0
numpy==1.24.3
qdrant-client==1.7.0
